Factor out db_connection_pool.ts
